+++ src/pykeen/datasets/temporal/base.py
@@ -7,7 +7,7 @@ import pathlib
 from typing import Any, Iterable, Mapping, Optional, Union
 
 from pykeen.datasets.base import LazyDataset
-from pykeen.triples.quadruples_factory import QuadruplesFactory
+from pykeen.triples import QuadruplesFactory
 
 __all__ = ["TemporalPathDataset"]
