+++ src/pykeen/datasets/temporal/small_sample/train.txt
@@ -10,4 +10,4 @@
 6	4	0	2000-01-07
 2	2	9	2000-01-09
 1	2	4	2000-01-11
-1   2   4   2000-01-12
+1	2	4	2000-01-12
