+++ src/pykeen/triples/__init__.py
@@ -3,8 +3,7 @@
 """Classes for creating and storing training data from triples."""
 
 from .instances import Instances, LCWAInstances, SLCWAInstances, LCWAQuadrupleInstances, QuadrupleInstances, SLCWAQuadrupleInstances
-from .quadruples_factory import QuadruplesFactory
-from .triples_factory import AnyTriples, CoreTriplesFactory, KGInfo, TriplesFactory, get_mapped_triples
+from .triples_factory import AnyTriples, CoreTriplesFactory, KGInfo, TriplesFactory, QuadruplesFactory, get_mapped_triples
 from .triples_numeric_literals_factory import TriplesNumericLiteralsFactory
 
 __all__ = [
